Refactoring DatabaseManager -> ResourceManager
